refactor: extract the generic parts of the IstioInfo* classes
This takes the concrete IstioInfoProvider and Requirer classes and rewrites the in terms of a generic Receiver and Provider base class.  We then implement the IstioInfo* classes as subclasses that provide their own schema.

The future plan for this change is to extract the Requirer and Sender classes into a separate repo