feat: implement a collect_unit_status handler for status setting (#43)
* feat: implement a collect_unit_status handler for status setting

Refactors the status handling of the charm to use the collect_unit_status event for status setting.  In this PR, the change is effectively non-existent because the charm presently has minimal status setting, but a follow-on PR will implement other checks that may block the charm and should be in collect-status.