ci: fix how test_istio_metadata_lib.py checks relation data (#56)
Previously, we accessed .local_app_data of the relation object we created.  This worked because ops mutated the relation object we passed to it, but this behaviour was a bug that has since been patched in ops.

This PR refactors to use the intended workflow, which is pulling relation data from a generated State